perf: (Backend) - Add Table: DetailedDeviceLocation many to one DeviceLocation - Add feature GetDeviceLocations by DataUnitId

+++ backend/app/Http/Queries/DeviceLocations/DeleteDeviceLocationCommand.php
@@ -3,6 +3,7 @@
 namespace App\Http\Queries\DeviceLocations;
 
 use App\Models\DeviceLocation;
+use App\Models\DetailedDeviceLocation;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class DeleteDeviceLocationCommand
@@ -18,6 +19,13 @@ class DeleteDeviceLocationCommand
     {
         $deviceLocation = DeviceLocation::findOrFail($id); // Fetch the DeviceLocation by ID
 
+        // Check if the DeviceLocation is being used in the DetailedDeviceLocation table
+        $detailedDeviceLocationCount = DetailedDeviceLocation::where('DeviceLocationId', $id)->count();
+
+        if ($detailedDeviceLocationCount > 0) {
+            return response()->json(['message' => 'DeviceLocation is being used in DetailedDeviceLocation'], 409);
+        }
+        
         $deviceLocation->delete(); // Soft delete the DeviceLocation
 
         return response()->json(['message' => 'DeviceLocation deleted successfully'], 200);
